Update Psalm to v2. Support contracts are available.

+++ src/Filter/Special/CreditCardNumberFilter.php
@@ -43,7 +43,7 @@ class CreditCardNumberFilter extends StringFilter
         $l = \count($split);
         for ($i = 0; $i < $l; ++$i) {
             /** @var int $n */
-            $n = $split[$l - $i - 1] << ($i & 1);
+            $n = (int) ($split[$l - $i - 1]) << ($i & 1);
             if ($n > 9) {
                 $n = ((int) ($n / 10)) + ($n % 10);
             }
